Refactor: Break apart tests/test_api.py into two separate files - test_extract_text_success
